Utilize Honeybadger::Sender as a base namespace for additional senders.
