+++ honeybadger.gemspec
@@ -97,8 +97,8 @@ Gem::Specification.new do |s|
     lib/honeybadger/railtie.rb
     lib/honeybadger/rake_handler.rb
     lib/honeybadger/sender.rb
-    lib/honeybadger/development_sender.rb
-    lib/honeybadger/test_sender.rb
+    lib/honeybadger/sender/development.rb
+    lib/honeybadger/sender/test.rb
     lib/honeybadger/shared_tasks.rb
     lib/honeybadger/tasks.rb
     lib/honeybadger_tasks.rb
