+++ lib/honeybadger/sender/development.rb
@@ -1,0 +1,20 @@
+module Honeybadger
+  class Sender
+    class Development
+
+      def initialize(options = {})
+      end
+
+      # Public: Sends the notice data off to no where.
+      #
+      # notice - The notice data to be sent (Hash or JSON string)
+      #
+      # Returns nil
+      def send_to_honeybadger(notice)
+        message = notice.is_a?(String) ? notice : notice.error_message
+        Honeybadger.write_verbose_log(message, :debug)
+      end
+
+    end
+  end
+end
